encoder mapped to drive + fixes to bipolar encoders

+++ csd2c/Eindopdracht/hardClip.cpp
@@ -11,7 +11,7 @@ Hardclip::~Hardclip()
 
 void Hardclip::tick()
 {
-  drive = std::max(drive, 0); //Protection against negative numbers
+  drive = std::max(drive, (float)0.0); //Protection against negative numbers
   float x1 = sample_dry * drive;
 
   if(x1 > 1){
@@ -34,7 +34,7 @@ void Hardclip::write(float input)
   sample_dry = input;
 }
 
-void Hardclip::setDrive(int drive_in)
+void Hardclip::setDrive(float drive_in)
 {
   this->drive = drive_in;
 }
